# Avoid `c()` calls when only `x` rows are required